Export files moved under document/ folder
Resolves #75
