+++ src/document/export/formatRTF.js
@@ -4,7 +4,7 @@
 //
 // ****************************************************************************
 
-import {getHeader} from "../../document/head"
+import {getHeader} from "../head"
 
 //-----------------------------------------------------------------------------
