+++ src/util/generic.js
@@ -7,6 +7,11 @@
 //*****************************************************************************
 
 //-----------------------------------------------------------------------------
+
+export const isEmpty = x => !x;
+export const isNotEmpty = x => !!x;
+
+//-----------------------------------------------------------------------------
 // Number formatting
 //-----------------------------------------------------------------------------
